Upd itemLic api
